+++ qml/SectionCard.qml
@@ -50,7 +50,6 @@ Kirigami.AbstractCard {
                     Layout.fillWidth: true
                     visible: root.description.length > 0
                     text: root.description
-                    color: Kirigami.Theme.disabledTextColor
                     wrapMode: Text.WordWrap
                 }
             }